modified:   app/Http/Controllers/AnthropometryController.php modified:   app/Http/Controllers/CoupleController.php new file:   app/Http/Controllers/Deleted/DeleteCoupleController.php new file:   app/Http/Controllers/Deleted/DeleteFamilyController.php new file:   app/Http/Controllers/Deleted/DeletePersonController.php new file:   app/Http/Controllers/Deleted/DeletePosyanduController.php new file:   app/Http/Controllers/Deleted/DeletePregnancyController.php new file:   app/Http/Controllers/Deleted/DeletePuerperalController.php deleted:    app/Http/Controllers/KeluargaBerencana/MonthlyReport.php modified:   app/Http/Controllers/PersonController.php modified:   app/Http/Controllers/PosyanduController.php modified:   app/Http/Controllers/PregnancyController.php modified:   app/Http/Controllers/PuerperalController.php modified:   app/Models/Anthropometry.php modified:   app/Models/Couple.php modified:   app/Models/Family.php modified:   app/Models/Neonatus.php modified:   app/Models/Person.php modified:   app/Models/Posyandu.php modified:   app/Models/Pregnancy.php modified:   app/Models/PrenatalClass.php modified:   app/Models/Puerperal.php modified:   app/Models/PuerperalClass.php modified:   app/Services/AnthropometryService.php modified:   app/Services/CoupleService.php modified:   app/Services/FamilyService.php modified:   app/Services/KeluargaBerencanaService.php modified:   app/Services/PersonService.php modified:   app/Services/PosyanduService.php modified:   app/Services/PregnancyService.php modified:   app/Services/PuerperalService.php new file:   database/migrations/2021_11_04_223208_add_deleted_at_to_posyandu_table.php new file:   database/migrations/2021_11_05_090219_add_deleted_at_to_puerperals_table.php new file:   database/migrations/2021_11_05_133431_add_deleted_at_to_people_table.php deleted:    public/windmill/public/buttons.html deleted:    public/windmill/public/cards.html deleted:    public/windmill/public/charts.html deleted:    public/windmill/public/forms.html deleted:    public/windmill/public/index.html deleted:    public/windmill/public/modals.html deleted:    public/windmill/public/pages/404.html deleted:    public/windmill/public/pages/blank.html deleted:    public/windmill/public/pages/create-account.html deleted:    public/windmill/public/pages/forgot-password.html deleted:    public/windmill/public/pages/login.html deleted:    public/windmill/public/tables.html modified:   resources/views/auth/forgot-password.blade.php modified:   resources/views/components/user/create-user.blade.php new file:   resources/views/deleted-models/deleted-couples.blade.php new file:   resources/views/deleted-models/deleted-families.blade.php new file:   resources/views/deleted-models/deleted-people.blade.php new file:   resources/views/deleted-models/deleted-posyandu.blade.php new file:   resources/views/deleted-models/deleted-pregnancies.blade.php new file:   resources/views/deleted-models/deleted-puerperals.blade.php deleted:    resources/views/imports.blade.php modified:   resources/views/layouts/sidebar.blade.php modified:   resources/views/people/create.blade.php modified:   resources/views/people/edit.blade.php modified:   resources/views/people/index.blade.php modified:   resources/views/posyandu/anthropometries/create.blade.php modified:   resources/views/posyandu/anthropometries/edit.blade.php modified:   resources/views/posyandu/edit.blade.php deleted:    resources/views/posyandu/neonatuses/index.blade.php modified:   resources/views/pregnancies/edit.blade.php modified:   resources/views/pregnancies/show.blade.php modified:   resources/views/puerperals/edit.blade.php modified:   resources/views/puerperals/show.blade.php deleted:    resources/views/template/index.blade.php modified:   resources/views/users/index.blade.php modified:   routes/auth.php modified:   routes/web.php

+++ app/Services/CoupleService.php
@@ -74,22 +74,64 @@ class CoupleService
 
         $couple->update($attributes);
     }
+
+    public function getDeletedCouples()
+    {
+        return Couple::onlyTrashed()->get();
+    }
     
     /**
      * delete couple dengan softdelete
      *
      * @return void
      */
-    public function delete($request, $couple)
+    public function softDelete($request, $couple)
     {
         // ubah status kawin istri menjadi cerai
         $istri = $couple->wife;
+        $suami = $couple->husband;
         $istri->update(['marital_status_id' => $request->marital_status_id]);
+        $suami->update(['marital_status_id' => $request->marital_status_id]);
 
         # hapus dulu KB datanya
-        $couple->keluargaBerencana()->delete();
-        # hapus couplenya
+        if (isset($couple->keluargaBerencana)) {
+            $couple->keluargaBerencana()->delete();
+        }
+
         $couple->delete();
+    }
+
+    public function forceDelete($couple)
+    {
+        $couple = Couple::withTrashed()->find($couple);
+
+        # hapus dulu KB datanya
+        if (isset($couple->keluargaBerencana)) {
+            $couple->keluargaBerencana()->forceDelete();
+        }
+
+        $couple->forceDelete();
+    }
+
+    public function restore($couple)
+    {
+        $couple = Couple::withTrashed()->find($couple);
+        
+        $wife = $couple->wife;
+        // cek apakah istri sudah punya suami lagi, krn satu istri hanya boleh satu suami
+        if (isset($wife->husband)) {
+            return false;
+        } else {
+            if (isset($couple->keluargaBerencana)) {
+                $couple->keluargaBerencana()->restore();
+            }
+    
+            $couple->restore();
+            $couple->wife->update(['marital_status_id' => 2]);
+            $couple->husband->update(['marital_status_id' => 2]);
+
+            return true;
+        }
     }
     
     /**
